+++ spec/module_spec.lua
@@ -11,6 +11,7 @@ describe('module', function()
   end)
 
   describe('.call', function()
+    after_each(function() module.flush() end)
 
     it('returns', function()
       package.loaded['foobar'] = { phase = function() return 'foobar' end }
@@ -23,5 +24,19 @@ describe('module', function()
       assert.falsy(err)
     end)
 
+  end)
+
+  describe(':require', function()
+    it('returns a module', function()
+      local foobar = { _VERSION = '1.1', _NAME = 'Foo Bar' }
+      package.loaded['foobar'] = foobar
+
+      local m = module.new('foobar')
+
+      local mod, err = m:require()
+
+      assert.equal(mod, foobar)
+      assert.falsy(err)
+    end)
   end)
 end)